Pre-specified transcription Factors from the UToronto dataset

This page contains the official list of human TFs from Lambert et al (PMID:29425488), along with all associated information. Candidate proteins were manually examined by a panel of experts based on available data. Proteins with experimentally demonstrated DNA binding specificity were considered TFs. Other proteins, such as co-factors and RNA binding proteins, were classified as non-TFs. All proteins (both TFs and non-TFs) are contained in the database, along with the associated evidence.

Usage

UTORONTO_TFs

Format

https://humantfs.ccbr.utoronto.ca/download.php