Wrapper for CAMERA-PR

A wrapper for cameraPR.matrix that performs pre-ranked Correlation Adjusted MEan RAnk molecular signature analysis of differential analysis results.

Usage

run_cameraPR(
  DA_list = NULL,
  selected_omes = c("transcript-rna-seq", "prot-pr", "prot-ph", "prot-ol", "metab"),
  selected_tissues = "all",
  database = setdiff(names(MotrpacHumanPreSuspensionAnalysis::MOLECULAR_SIGNATURES),
    "PTMSIGDB"),
  path_to_gmt = NULL,
  min_size = 5L,
  overlap_cutoff = 0.7
)

Arguments

DA_list

list; a named list of data.frame objects, each containing differential analysis results for a specific tissue/assay combination. The list must be nested, with tissues at the top level and omes within tissues. If NULL (default), the differential analysis results will be generated with load_differential_analysis. Unless wishing to analyze DA results that are not in MotrpacHumanPreSuspensionAnalysis, this should remain NULL.

selected_omes

character; one or more character strings selected from the following options: "transcript-rna-seq", "prot-pr", "prot-ph", and "metab" (all metabolomics platforms). Passed to load_differential_analysis.

selected_tissues

character; passed to load_differential_analysis. One or more of the following: "all", "muscle", "adipose", or "blood".

database

character; one or more names specifying the database(s) to test. Options are (case insensitive) "BIOCARTA", "KEGG_MEDICUS", "PID", "REACTOME", "WP" (WikiPathways database), "GOBP", "GOCC", "GOMF", "MITOCARTA" (MitoCarta3.0 database), "PSP" (PhosphoSitePlus kinases; only valid when selected_omes contains "prot-ph"), or "REFMET" (RefMet chemical subclasses; only valid when selected_omes contains "metab"). See MOLECULAR_SIGNATURES for details.

path_to_gmt

character; (optional) path to one or more GMT files. Passed to TMSig::readGMT. If provided, database is ignored.

min_size

integer; the minimum set size for testing.

overlap_cutoff

numeric; the minimum proportion of genes in each set that must appear in a given dataset. Used to pre-filter sets. Does not affect "metab" or "prot-ph" results. This will always be 0.1 for "prot-ol" results.

Value

An object of class data.frame with the following columns:

tissue

factor; the tissue.

html
\item{\code{assay}}{factor; the omics assay.} \item{\code{contrast_type}}{factor; the type of contrast.} \item{\code{contrast}}{factor; the contrast of interest.} \item{\code{contrast_short}}{factor; shortened contrasts.} \item{\code{collection}}{factor; the broad molecular signature collection. Only included when \code{path_to_gmt} is \code{NULL}. See \code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}} for details.} \item{\code{database}}{factor; the molecular signature database. Only included when \code{path_to_gmt} is \code{NULL}. See \code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}} for details.} \item{\code{set_id}}{character; a unique ID for the molecular signature. See \code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}} for details.} \item{\code{set}}{character; the molecular signature being tested. For global proteomics and transcriptomics, these are gene sets. For phosphoproteomics, these are kinase sets.} \item{\code{set_short}}{character; a shortened version of \code{set}. Only included when \code{path_to_gmt} is \code{NULL}. See \code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}} for details.} \item{\code{set_size}}{integer; the number of molecules in the set that were present in the DA results for that specific tissue/assay combination.} \item{\code{set_size_DB}}{integer; the number of molecules in the set, as defined in the GMT file.} \item{\code{size_ratio}}{numeric; the ratio of \code{set_size} to \code{set_size_DB}, rounded to the nearest thousandth. A measure of confidence that the gene set being tested is correctly described by the entry in the \code{set} column. While smaller values do not necessarily indicate that the results are unreliable, terms from the gene set databases should be treated with caution.} \item{\code{direction}}{factor; the direction of change: \code{"Up"} or \code{"Down"}.} \item{\code{t}}{numeric; the two-sample t-statistic.} \item{\code{df}}{integer; the available degrees of freedom.} \item{\code{z.std}}{numeric; standard Normal equivalents of \code{t}.} \item{\code{p_value}}{numeric; the two-sided p-value.} \item{\code{adj_p_value}}{numeric; the BH-adjusted p-value. P-values are adjusted within each combination of tissue, assay, contrast, and collection.} html

See also

Author

Tyler Sagendorf

Examples

if (FALSE) { # \dontrun{
# Test Reactome and Gene Ontology Biological Processes
# databases on muscle global proteomics.
x1 <- run_cameraPR(selected_omes = "prot-pr",
                   selected_tissues = "muscle",
                   database = c("GOBP", "REACTOME"))

# Test all omes using all molecular signature databases
x2 <- run_cameraPR()
} # }